Adapt 1.20.1 code (procedure full)

+++ src/main/java/com/pla/annoyingvillagers/procedures/EnchantBedItemOnRightClickBlockProcedure.java
@@ -14,7 +14,7 @@ public class EnchantBedItemOnRightClickBlockProcedure {
 
     public static void execute(LevelAccessor levelaccessor, double d0, double d1, double d2, Entity entity) {
         if (entity != null) {
-            levelaccessor.setBlock(new BlockPos(d0, d1 + 1.0D, d2), ((Block) AnnoyingVillagersModBlocks.ENCHANT_BED.get()).defaultBlockState(), 3);
+            levelaccessor.setBlock(new BlockPos((int) d0, (int) d1 + 1, (int) d2), ((Block) AnnoyingVillagersModBlocks.ENCHANT_BED.get()).defaultBlockState(), 3);
             if (entity instanceof Player) {
                 Player player = (Player) entity;
                 ItemStack itemstack = new ItemStack((ItemLike) AnnoyingVillagersModItems.ENCHANT_BED_ITEM.get());
